Break import cycle (#273) (#274)
* Move ``ensure`` definition from utils to exceptions

* Use correct import for ``ensure``

* Test relocated ``ensure`` asserts replacer

* Move ``ensure`` tests to a file of their own.
