+++ src/nacl/bindings/crypto_shorthash.py
@@ -15,7 +15,6 @@
 from __future__ import absolute_import, division, print_function
 
 import nacl.exceptions as exc
-from nacl import utils
 from nacl._sodium import ffi, lib
 
 BYTES = lib.crypto_shorthash_siphash24_bytes()
@@ -37,5 +36,5 @@ def crypto_shorthash_siphash24(data, key):
     digest = ffi.new("unsigned char[]", BYTES)
     rc = lib.crypto_shorthash_siphash24(digest, data, len(data), key)
 
-    utils.ensure(rc == 0, raising=exc.RuntimeError)
+    exc.ensure(rc == 0, raising=exc.RuntimeError)
     return ffi.buffer(digest, BYTES)[:]
