+++ src/nacl/bindings/crypto_sign.py
@@ -16,7 +16,7 @@ from __future__ import absolute_import, division, print_function
 
 from nacl import exceptions as exc
 from nacl._sodium import ffi, lib
-from nacl.utils import ensure
+from nacl.exceptions import ensure
 
 
 crypto_sign_BYTES = lib.crypto_sign_bytes()
